Remove TextureTestMixin
Using inheritance this way ends up causing issues
like the diamond pattern where 2 downstream classes
want to both inherit from multiple "Mixins".

All of these functions run just fine as standalone
functions. The issue of discovery was brought up as one
reason to use Mixin but we can use the `import * as name`
pattern to pull in groups of related functions.

Note: I plan to try fix ValidationTest similarly, separating
out all of its helpers in validation_test_utils.ts. It's
duplicating several hundred lines of code right now because
of the diamond pattern issue.
