+++ src/webgpu/api/operation/sampling/anisotropy.spec.ts
@@ -12,7 +12,8 @@ things. If there are no guarantees we can issue warnings instead of failures. Id
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
 import { assert } from '../../../../common/util/util.js';
-import { AllFeaturesMaxLimitsGPUTest, TextureTestMixin } from '../../../gpu_test.js';
+import { AllFeaturesMaxLimitsGPUTest } from '../../../gpu_test.js';
+import * as ttu from '../../../texture_test_utils.js';
 import { checkElementsEqual } from '../../../util/check_contents.js';
 import { TexelView } from '../../../util/texture/texel_view.js';
 import { PerPixelComparison } from '../../../util/texture/texture_ok.js';
@@ -161,7 +162,7 @@ class SamplerAnisotropicFilteringSlantedPlaneTest extends AllFeaturesMaxLimitsGP
   }
 }
 
-export const g = makeTestGroup(TextureTestMixin(SamplerAnisotropicFilteringSlantedPlaneTest));
+export const g = makeTestGroup(SamplerAnisotropicFilteringSlantedPlaneTest);
 
 g.test('anisotropic_filter_checkerboard')
   .desc(
@@ -286,7 +287,8 @@ g.test('anisotropic_filter_mipmap_color')
     },
   ])
   .fn(t => {
-    const texture = t.createTextureFromTexelViewsMultipleMipmaps(
+    const texture = ttu.createTextureFromTexelViewsMultipleMipmaps(
+      t,
       colors.map(value => TexelView.fromTexelsAsBytes(kTextureFormat, _coords => value)),
       { size: [4, 4, 1], usage: GPUTextureUsage.COPY_DST | GPUTextureUsage.TEXTURE_BINDING }
     );
@@ -321,5 +323,9 @@ g.test('anisotropic_filter_mipmap_color')
         );
       }
     }
-    t.expectSinglePixelComparisonsAreOkInTexture({ texture: colorAttachment }, pixelComparisons);
+    ttu.expectSinglePixelComparisonsAreOkInTexture(
+      t,
+      { texture: colorAttachment },
+      pixelComparisons
+    );
   });
