+++ src/webgpu/shader/execution/expression/call/builtin/subgroup_util.ts
@@ -1,14 +1,15 @@
 import { assert, iterRange, unreachable } from '../../../../../../common/util/util.js';
 import { Float16Array } from '../../../../../../external/petamoriken/float16/float16.js';
 import { getBlockInfoForTextureFormat } from '../../../../../format_info.js';
-import { AllFeaturesMaxLimitsGPUTest, GPUTest, TextureTestMixin } from '../../../../../gpu_test.js';
+import { AllFeaturesMaxLimitsGPUTest, GPUTest } from '../../../../../gpu_test.js';
+import * as ttu from '../../../../../texture_test_utils.js';
 import { kBit } from '../../../../../util/constants.js';
 import { Type, VectorType, scalarTypeOf } from '../../../../../util/conversion.js';
 import { FPInterval } from '../../../../../util/floating_point.js';
 import { sparseScalarF16Range, sparseScalarF32Range, align } from '../../../../../util/math.js';
 import { PRNG } from '../../../../../util/prng.js';
 
-export class SubgroupTest extends TextureTestMixin(AllFeaturesMaxLimitsGPUTest) {}
+export class SubgroupTest extends AllFeaturesMaxLimitsGPUTest {}
 
 export const kNumCases = 1000;
 export const kStride = 128;
@@ -574,7 +575,7 @@ fn vsMain(@builtin(vertex_index) index : u32) -> @builtin(position) vec4f {
   pass.end();
   t.queue.submit([encoder.finish()]);
 
-  const copyBuffer = t.copyWholeTextureToNewBufferSimple(framebuffer, 0);
+  const copyBuffer = ttu.copyWholeTextureToNewBufferSimple(t, framebuffer, 0);
   const readback = await t.readGPUBufferRangeTyped(copyBuffer, {
     srcByteOffset: 0,
     type: Uint32Array,
