+++ src/webgpu/shader/execution/shader_io/vertex_builtins.spec.ts
@@ -4,9 +4,9 @@ export const description = `Test vertex shader builtin variables
 `;
 
 import { makeTestGroup } from '../../../../common/framework/test_group.js';
-import { AllFeaturesMaxLimitsGPUTest, TextureTestMixin } from '../../../gpu_test.js';
+import { AllFeaturesMaxLimitsGPUTest } from '../../../gpu_test.js';
 
-class VertexBuiltinTest extends TextureTestMixin(AllFeaturesMaxLimitsGPUTest) {}
+class VertexBuiltinTest extends AllFeaturesMaxLimitsGPUTest {}
 
 export const g = makeTestGroup(VertexBuiltinTest);
