+++ src/webgpu/web_platform/copyToTexture/video.spec.ts
@@ -8,7 +8,8 @@ copyToTexture with HTMLVideoElement and VideoFrame.
 `;
 
 import { makeTestGroup } from '../../../common/framework/test_group.js';
-import { AllFeaturesMaxLimitsGPUTest, TextureTestMixin } from '../../gpu_test.js';
+import { AllFeaturesMaxLimitsGPUTest } from '../../gpu_test.js';
+import * as ttu from '../../texture_test_utils.js';
 import {
   startPlayingAndWaitForVideo,
   getVideoElement,
@@ -22,7 +23,7 @@ import {
 
 const kFormat = 'rgba8unorm';
 
-export const g = makeTestGroup(TextureTestMixin(AllFeaturesMaxLimitsGPUTest));
+export const g = makeTestGroup(AllFeaturesMaxLimitsGPUTest);
 
 g.test('copy_from_video')
   .desc(
@@ -101,7 +102,7 @@ It creates HTMLVideoElement with videos under Resource folder.
       const expect = kVideoInfo[videoName].display;
 
       if (srcDoFlipYDuringCopy) {
-        t.expectSinglePixelComparisonsAreOkInTexture({ texture: dstTexture }, [
+        ttu.expectSinglePixelComparisonsAreOkInTexture(t, { texture: dstTexture }, [
           // Flipped top-left.
           {
             coord: { x: width * 0.25, y: height * 0.25 },
@@ -124,7 +125,7 @@ It creates HTMLVideoElement with videos under Resource folder.
           },
         ]);
       } else {
-        t.expectSinglePixelComparisonsAreOkInTexture({ texture: dstTexture }, [
+        ttu.expectSinglePixelComparisonsAreOkInTexture(t, { texture: dstTexture }, [
           // Top-left.
           {
             coord: { x: width * 0.25, y: height * 0.25 },
